examples/vertexai/requirements-dev.txt
google-cloud-aiplatform
google-cloud-bigquery
google-cloud-pipeline-components<2
kfp<2
numpy
pandas
pyarrow
scikit-learn
sf-hamilton[visualization]
